Add ClientID filtering

+++ bevy_hookup_core/src/client_id.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Reflect, Resource, Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
 #[reflect(Default)]
-pub struct ClientId(u64);
+pub struct ClientId(u32);
 
 impl Default for ClientId {
     fn default() -> Self {
